database contentprovider design completed

+++ app/src/main/java/lordsomen/android/com/letsbake/data/BakingAppDao.java
@@ -20,7 +20,7 @@ public interface BakingAppDao {
     Cursor selectAll();
 
     @Query("SELECT * FROM baking_data WHERE id =:id")
-    Cursor selectById(int id);
+    Cursor selectById(long id);
 
     @Query("SELECT COUNT(*) FROM baking_data")
     int count();
